examples/chart-generator/requirements.txt
praisonai-svc>=1.2.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
numpy>=1.24.0



